better restrict int index size

+++ src/board.nim
@@ -4,7 +4,7 @@
 
 import ./datatypes
 
-proc neighbors*[N: static[int]](idx: Point[N]): array[4, Point[N]] {.inline.}=
+proc neighbors*[N: static[int16]](idx: Point[N]): array[4, Point[N]] {.inline.}=
   [idx-1, idx+1, idx - (N+2), idx + (N+2)]
